Functions/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
gremlinpython==3.4.8
azure.cosmos
azure-eventhub
azure-eventhub-checkpointstoreblob-aio
